+++ src/main/java/org/opentripplanner/model/Entrance.java
@@ -13,7 +13,7 @@ public final class Entrance extends StationElement {
     FeedScopedId id,
     I18NString name,
     String code,
-    String description,
+    I18NString description,
     WgsCoordinate coordinate,
     WheelchairAccessibility wheelchairAccessibility,
     StopLevel level