+++ docs/Changelog.md
@@ -60,6 +60,7 @@ based on merged pull requests. Search GitHub issues and pull requests for smalle
 - Do not link cars to stop vertices in routing [#4166](https://github.com/opentripplanner/OpenTripPlanner/pull/4166)
 - Transit model - Part 3 [#4176](https://github.com/opentripplanner/OpenTripPlanner/pull/4176)
 - Add Siri realtime occupancy info [#4180](https://github.com/opentripplanner/OpenTripPlanner/pull/4180)
+- Add gtfs stop description translations [#4158](https://github.com/opentripplanner/OpenTripPlanner/pull/4158)
   [](AUTOMATIC_CHANGELOG_PLACEHOLDER_DO_NOT_REMOVE)
 
 ## 2.1.0 (2022-03-17)